Move button component to new file
